FlashImage.py: Avoid Device or resource busy
<  5224.5829>   File "/usr/lib/enigma2/python/Components/ActionMap.py", line 188, in action
<  5224.5838>   File "/usr/lib/enigma2/python/Screens/FlashImage.py", line 463, in cancel
<  5224.5858>     rmdir(self.tmp_dir)
<  5224.5863> OSError: [Errno 16] Device or resource busy: '/tmp/MultibootSelectionzhc8zg94'
<  5224.5865> [ePyObject] (CallObject(<bound method ActionMap.action of <Components.ActionMap.ActionMap object at 0xad3161d8>>,('OkCancelActions', 'cancel')) failed)

tmp dir is very sensitive.
If we enter the multiboot selector and in turn open the temporary folder with an FTP client, the resource is already active and the device is in use, this means that we cannot reboot with another image-slot, until we exit screen MultibootSelector and let's repeat the sequence again creating a new temporary folder.
If this folder is in use and could not be deleted, if we try to reboot into another image-slot, we get a loop.
So let's avoid these situations by making sure the temp folder doesn't exist to reboot to another slot or cancel the action.

+++ lib/python/Screens/FlashImage.py
@@ -459,9 +459,9 @@ class MultibootSelection(SelectImage):
 
 	def cancel(self, value=None):
 		Console().ePopen('umount %s' % self.tmp_dir)
-		if exists(self.tmp_dir):
+		if not ismount(self.tmp_dir) and exists(self.tmp_dir):
 			rmdir(self.tmp_dir)
-		if value == 2 and not isfile(join(self.tmp_dir, "STARTUP")):
+		if value == 2 and not exists(self.tmp_dir):
 			self.session.open(TryQuitMainloop, 2)
 		else:
 			self.close(value)
